+++ packages/core/src/modules/openid/refresh/types.ts
@@ -77,6 +77,6 @@ export interface IRefreshOrchestrator {
 }
 
 export enum OpenIDCustomNotificationType {
-  CredentialReplacementAvailable = 'CustomNotificationOpenIDCredential',
+  CredentialReplacementAvailable = 'CredentialRefresh',
   CredentialExpired = 'CredentialExpired',
 }